Added missing methods to lifecycle. Subscriptions are now handled on screen registry
