+++ src/core-examples/index.ts
@@ -1,6 +1,6 @@
 import * as Core from "core";
 import * as AppFormerComponents from "core-screens/AppFormerComponents";
-import * as Components from "core/Components";
+import * as ReactShorthandComponents from "core/react/Shorthands";
 
-Core.register({...Components,...AppFormerComponents});
+Core.register({...ReactShorthandComponents,...AppFormerComponents});
